New Evaluation Function
I created a new function or representation of the board, in order to help the computer understand how the game is played in slightly more depth. In the future, I will have to add a system for open rooks, passed pawns vs. blocked pawns, and King Safety.

The state needs to be added to the game.